数据格式、引入编码

+++ backend/ruoyi-system/src/main/java/com/ruoyi/system/domain/IcesProductStore.java
@@ -38,11 +38,11 @@ public class IcesProductStore extends BaseEntity {
     /**
      * 货位数量
      */
-    private Long prsSpace;
+    private Float prsSpace;
     /**
      * 空闲货位数量
      */
-    private Long prsFree;
+    private Float prsFree;
     /**
      * 已删除
      */
