+++ backend/ruoyi-system/src/main/java/com/ruoyi/system/domain/vo/IcesProductStoreVo.java
@@ -43,13 +43,13 @@ public class IcesProductStoreVo {
      * 货位数量
      */
     @ExcelProperty(value = "货位数量")
-    private Long prsSpace;
+    private Float prsSpace;
 
     /**
      * 空闲货位数量
      */
     @ExcelProperty(value = "空闲货位数量")
-    private Long prsFree;
+    private Float prsFree;
 
     /**
      * 已删除
